NavBar fully replaced
