+++ webapp/src/views/components/Nav.jsx
@@ -1,114 +1,167 @@
-import React, { useState, useEffect, useContext } from 'react';
-import { AppBar, Box, Container, IconButton, Toolbar, Tooltip, Button, Typography, Menu, MenuItem, Divider } from '@mui/material';
-import MenuIcon from '@mui/icons-material/Menu';
-import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown';
-import { Link } from 'react-router-dom';
+import AppBar from "@mui/material/AppBar";
+import {
+  Box, Button, Collapse,
+  Container, Divider,
+  Drawer,
+  IconButton,
+  List,
+  ListItemButton,
+  ListItemIcon,
+  ListItemText, Menu, MenuItem, Tooltip, Typography
+} from "@mui/material";
+import Toolbar from "@mui/material/Toolbar";
+import MenuIcon from "@mui/icons-material/Menu";
+import React, {useContext, useEffect, useState} from "react";
+import {Link} from "react-router-dom";
 import { ReactComponent as CustomIcon } from '../../media/logoS.svg';
-import { AuthContext } from '../context/AuthContext';
-import { ConfigContext } from '../context/ConfigContext';
-import MyAvatar from "./MyAvatar"
-import {LocaleContext} from "../context/LocaleContext"
+import {LocaleContext} from "../context/LocaleContext";
+import {AuthContext} from "../context/AuthContext";
+import PublicIcon from '@mui/icons-material/Public';
+import InfoIcon from '@mui/icons-material/Info';
+import WidgetsIcon from '@mui/icons-material/Widgets';
+import HomeIcon from '@mui/icons-material/Home';
+import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
+import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
+import LogoutIcon from '@mui/icons-material/Logout';
+import LoginIcon from '@mui/icons-material/Login';
+import PersonAddIcon from '@mui/icons-material/PersonAdd';
+import AccountCircleIcon from '@mui/icons-material/AccountCircle';
+import SwapHorizIcon from '@mui/icons-material/SwapHoriz';
+import MyAvatar from "./MyAvatar";
+import {ConfigContext} from "../context/ConfigContext";
 
 const pages = [
-    { code: 'home_home', link: '/home', logged: false},
-    { code: 'home_menu', link: '/menu', logged: true},
-    { code: 'home_ranking', link: '/ranking', logged: false},
-    { code: 'home_about', link: '/about', logged: false}
+  {code: 'home_home', link: '/home', logged: false, icon: HomeIcon},
+  {code: 'home_menu', link: '/menu', logged: true, icon: WidgetsIcon},
+  {code: 'home_ranking', link: '/ranking', logged: false, icon: PublicIcon},
+  {code: 'home_about', link: '/about', logged: false, icon: InfoIcon}
 ];
 
 const settings = [
-    { code: 'home_menu_account', link: '/account', logged: true },
-    { code: 'home_menu_login', link: '/login', logged: false },
-    { code: 'home_menu_sign_up', link: '/signup', logged: false },
-    { code: 'home_menu_logout', link: '/logout', logged: true }
+  {code: 'home_menu_account', link: '/account', logged: true, icon: AccountCircleIcon},
+  {code: 'home_menu_login', link: '/login', logged: false, icon: LoginIcon},
+  {code: 'home_menu_sign_up', link: '/signup', logged: false, icon: PersonAddIcon},
+  {code: 'home_menu_logout', link: '/logout', logged: true, icon: LogoutIcon}
 ];
 
 const locale = [
-  { code: "i18n_spanish", value: "es" },
-  { code: "i18n_english", value: "en" },
+  {code: "i18n_spanish", value: "es"},
+  {code: "i18n_english", value: "en"},
 ]
+
+const localeToText = (id) => locale.find(x => x.value === id).code
 
 const threshold = 899;
 
-const JordiButton = () => {
-  const { swapConfig } = useContext(ConfigContext);
-  const { t } = useContext(LocaleContext)
-
+const listGenerator = (data, t, isAuthenticated, setOpen, logout) => {
   return (
-      <Button
-        variant="contained"
-        color="secondary"
-        onClick={swapConfig}
-        sx={{
-          display: { xs: 'none', md: 'inline-flex' },
-          borderRadius: '50px',
-          fontWeight: 'bold',
-          textTransform: 'none'
-        }}
-      >
-        {t("nav_button_swap")}
-      </Button>
-  );
-};
-
-const NavMenu = ({handleCloseNavMenu}) => {
-  const { t } = useContext(LocaleContext)
-  const { isAuthenticated } = useContext(AuthContext)
-  return (
-    <Box sx={{ flexGrow: 1, display: { xs: 'none', md: 'flex' } }}>
-        {pages.map((page) => {
-          if (page.logged && !isAuthenticated()) return null
+    <List>
+      {
+        data.map(item => {
+          if(item.logged !== isAuthenticated) return null
+          const f = item.link === '/logout' ? () => {setOpen(false); logout()} : () => setOpen(false)
           return (
-            <Button
-              key={page.code}
-              onClick={handleCloseNavMenu}
-              component={Link}
-              sx={{ my: 2, color: 'white', display: 'block' }}
-              to={page.link}
-            >
-              {t(page.code)}
-            </Button>
+            <ListItemButton component={Link} to={item.link} onClick={f} key={item.code} disablePadding>
+              { item.icon && <ListItemIcon>
+                  <item.icon />
+                </ListItemIcon> }
+              <ListItemText primary={t(item.code)} />
+            </ListItemButton>
           )
-        })}
-    </Box>
+        })
+      }
+    </List>
   )
 }
 
 const NavIcon = ({width}) => (
   <Link style={{ margin: width > threshold ? '0.5rem' : '1rem auto' }} to="/home">
-      <CustomIcon />
+    <CustomIcon />
   </Link>
 );
 
-const DropDownMenu = ({handleCloseNavMenu, anchorElNav, setAnchorElNav}) => {
-  const handleOpenNavMenu = (event) => setAnchorElNav(event.currentTarget);
+const LeftDrawer = () => {
+  const { t, setLocale } = useContext(LocaleContext)
+  const { isAuthenticated, logout } = useContext(AuthContext)
   const { swapConfig } = useContext(ConfigContext);
-  const { t } = useContext(LocaleContext)
+  const [open, setOpen] = useState(false)
+  const {locale: l} = useContext(LocaleContext)
+  const [openLocale, setOpenLocale] = useState(false)
+
   return (
     <Box sx={{ flexGrow: { md: 1 }, display: { xs: 'flex', md: 'none' }, width: { xs: 'fit-content' } }}>
-        <IconButton size="large" onClick={handleOpenNavMenu} color="inherit">
-            <MenuIcon />
+      <IconButton size="large" color="inherit" onClick={() => setOpen(true)}>
+        <MenuIcon />
+      </IconButton>
+      <Drawer anchor="left" open={open} onClose={() => setOpen(false)}>
+        <Box>
+          {listGenerator(pages, t, isAuthenticated(), setOpen, logout)}
+        </Box>
+        <Divider />
+        <Box>
+          <List>
+            <ListItemButton onClick={swapConfig} disablePadding>
+              <ListItemIcon>
+                <SwapHorizIcon />
+              </ListItemIcon>
+              <ListItemText primary={t("nav_button_swap")} />
+            </ListItemButton>
+          </List>
+        </Box>
+        <Divider />
+        <Box>
+          <List>
+            <ListItemButton onClick={() => setOpenLocale(!openLocale)} disablePadding>
+              <ListItemIcon>
+                {openLocale ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
+              </ListItemIcon>
+              <ListItemText primary={t(localeToText(l))} />
+            </ListItemButton>
+            <Collapse in={openLocale} timeout="auto" unmountOnExit>
+              <List component="div" disablePadding>
+                {locale.map(l => {
+                  return (
+                    <ListItemButton key={l.code} onClick={() => {setOpenLocale(false); setOpen(false); setLocale(l.value)}} disablePadding>
+                      <ListItemText primary={t(l.code)} />
+                    </ListItemButton>
+                  )
+                })}
+              </List>
+            </Collapse>
+          </List>
+        </Box>
+      </Drawer>
+    </Box>
+  )
+}
+
+const UserIcon = () => {
+  const [open, setOpen] = useState(false)
+  const { isAuthenticated } = useContext(AuthContext)
+  const { t } = useContext(LocaleContext)
+
+  return (
+    <>
+      <Tooltip title="User profile">
+        <IconButton sx={{ p: 0 }} onClick={() => setOpen(true)} >
+          <MyAvatar />
         </IconButton>
-        <Menu
-          id="menu-appbar"
-          anchorEl={anchorElNav}
-          anchorOrigin={{ vertical: 'bottom', horizontal: 'left' }}
-          keepMounted
-          transformOrigin={{ vertical: 'top', horizontal: 'left' }}
-          open={Boolean(anchorElNav)}
-          onClose={handleCloseNavMenu}
-          sx={{ display: { xs: 'block', md: 'none' } }}
-        >
-          {pages.map((page) => (
-              <MenuItem key={page.code} onClick={handleCloseNavMenu} component={Link} to={page.link}>
-                  <Typography textAlign="center">{t(page.code)}</Typography>
-              </MenuItem>
-          ))}
-          <Divider />
-          <MenuItem onClick={() => {swapConfig();handleCloseNavMenu()}}>
-              <Typography textAlign="center">{t("nav_button_swap")}</Typography>
-          </MenuItem>
-        </Menu>
+      </Tooltip>
+      <Drawer anchor="right" open={open} onClose={() => setOpen(false)}>
+        <Box>
+          {listGenerator(settings, t, isAuthenticated(), setOpen)}
+        </Box>
+      </Drawer>
+    </>
+  )
+}
+
+const RightDrawer = () => {
+  return (
+    <Box sx={{ flexGrow: 0, display: "flex", flexFlow: "row", alignItems: "center", gap: "1rem" }}>
+      <UserIcon />
+      <LocaleMenu />
+      <JordiButton />
     </Box>
   )
 }
@@ -133,17 +186,15 @@ const LocaleMenu = () => {
         }}
         variant="outlined"
         color="secondary"
-        startIcon={<KeyboardArrowDownIcon />}
+        startIcon={open ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
       >
         {l}
       </Button>
       <Menu
+        sx={{ marginTop: '1rem' }}
         anchorEl={anchorEl}
         open={open}
         onClose={handleClose}
-        MenuListProps={{
-          'aria-labelledby': 'basic-button',
-        }}
       >
         {locale.map((l) =>
           <MenuItem
@@ -157,74 +208,72 @@ const LocaleMenu = () => {
   )
 }
 
-export default function Nav() {
+const JordiButton = () => {
+  const { swapConfig } = useContext(ConfigContext);
   const { t } = useContext(LocaleContext)
-  const { isAuthenticated, logout } = useContext(AuthContext)
-  const [anchorElNav, setAnchorElNav] = useState(null);
-  const [anchorElUser, setAnchorElUser] = useState(null);
+
+  return (
+    <Button
+      variant="contained"
+      color="secondary"
+      onClick={swapConfig}
+      sx={{
+        display: { xs: 'none', md: 'inline-flex' },
+        borderRadius: '50px',
+        fontWeight: 'bold',
+        textTransform: 'none'
+      }}
+    >
+      {t("nav_button_swap")}
+    </Button>
+  );
+};
+
+const Links = () => {
+  const { t } = useContext(LocaleContext)
+  const { isAuthenticated } = useContext(AuthContext)
+
+  return (
+    <Box sx={{ flexGrow: 1, display: { xs: 'none', md: 'flex' } }}>
+      {pages.map((page) => {
+        if (page.logged && !isAuthenticated()) return null
+        return (
+          <Button
+            key={page.code}
+            component={Link}
+            sx={{ my: 2, color: 'white', display: 'block' }}
+            to={page.link}
+          >
+            {t(page.code)}
+          </Button>
+        )
+      })}
+    </Box>
+  )
+}
+
+const Nav = () => {
   const [width, setWidth] = useState(window.innerWidth);
 
-  const handleOpenUserMenu = (event) => setAnchorElUser(event.currentTarget);
-  const handleCloseNavMenu = () => setAnchorElNav(null);
-  const handleCloseUserMenu = () => setAnchorElUser(null);
   const handleWindowResize = () => {
     const handleResizeWindow = () => setWidth(window.innerWidth);
     window.addEventListener('resize', handleResizeWindow);
     return () => window.removeEventListener('resize', handleResizeWindow);
   }
-
   useEffect(handleWindowResize, [])
-
-  const generateMenuItems = () => {
-    return settings.map((setting) => {
-      if (isAuthenticated() !== setting.logged) return null
-      return (
-        <MenuItem
-          key={setting.code}
-          onClick={setting.link === '/logout' ? () => {handleCloseUserMenu(); logout()} : handleCloseUserMenu}
-          component={Link}
-          to={setting.link === '/logout' ? "/" : setting.link}>
-          <Typography textAlign="center">{t(setting.code)}</Typography>
-        </MenuItem>
-      )
-    })
-  }
 
   return (
     <AppBar position="sticky">
       <Container maxWidth="xl">
         <Toolbar disableGutters>
-          <DropDownMenu
-            handleCloseNavMenu={handleCloseNavMenu}
-            anchorElNav={anchorElNav}
-            setAnchorElNav={setAnchorElNav}
-          />
+          <LeftDrawer />
           <NavIcon width={width} />
-          <NavMenu handleCloseNavMenu={handleCloseNavMenu} />
-          <Box sx={{ flexGrow: 0, display: "flex", flexFlow: "row", alignItems: "center", gap: "1rem" }}>
-            <Tooltip title="User profile">
-              <IconButton onClick={handleOpenUserMenu} sx={{ p: 0 }}>
-                <MyAvatar />
-              </IconButton>
-            </Tooltip>
-            <Menu
-              sx={{ mt: '45px' }}
-              id="menu-appbar"
-              anchorEl={anchorElUser}
-              anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
-              keepMounted
-              transformOrigin={{ vertical: 'top', horizontal: 'right' }}
-              open={Boolean(anchorElUser)}
-              onClose={handleCloseUserMenu}
-            >
-              {generateMenuItems()}
-            </Menu>
-            <LocaleMenu />
-            <JordiButton />
-          </Box>
+          <Links />
+          <RightDrawer />
         </Toolbar>
       </Container>
     </AppBar>
-  );
-  
-};
+  )
+}
+
+export default Nav
